infra/gcp/function/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-run==0.*
